Simplify device sync build-safe projections
