+++ apps/web/src/lib/device-sync/public-ingress-service.ts
@@ -13,11 +13,14 @@ import {
 
 import type { HostedDeviceSyncControlPlaneContext } from "./control-plane-context";
 import {
+  toHostedBrowserDeviceSyncConnectionSource,
+  type HostedBrowserDeviceSyncConnectionSource,
+} from "./browser-connection-source";
+import {
   createHostedBrowserConnectionId,
   toHostedBrowserDeviceSyncConnection,
   type HostedBrowserDeviceSyncConnection,
 } from "./public-connection";
-import type { HostedDeviceConnectionSource } from "./prisma-store";
 import {
   disconnectHostedDeviceSyncConnection,
   handleHostedDeviceSyncConnectionEstablished,
@@ -257,64 +260,4 @@ export class HostedDeviceSyncPublicIngressService {
     });
   }
 
-}
-
-export interface HostedBrowserDeviceSyncConnectionSource {
-  connectionId: string;
-  firstSeenAt: string;
-  lastSeenAt: string;
-  requiresReconnect?: boolean;
-  resourceCount: number;
-  sourceProviderSlug: string;
-  status: HostedDeviceConnectionSource["status"];
-}
-
-function toHostedBrowserDeviceSyncConnectionSource(
-  source: HostedDeviceConnectionSource,
-  browserConnectionId: string,
-): HostedBrowserDeviceSyncConnectionSource {
-  return {
-    connectionId: browserConnectionId,
-    firstSeenAt: source.firstSeenAt,
-    lastSeenAt: source.lastSeenAt,
-    ...(requiresConnectionSourceReconnect(source) ? { requiresReconnect: true } : {}),
-    resourceCount: countSourceResources(source.resourceAvailabilitySummary),
-    sourceProviderSlug: source.sourceProviderSlug,
-    status: source.status,
-  };
-}
-
-const CONNECTION_SOURCE_SUMMARY_METADATA_KEYS = new Set([
-  "sourceInstanceKeyFallback",
-]);
-const CONNECTION_SOURCE_RECONNECT_ERROR_CODES = new Set(["TOKEN_REFRESH_FAILED"]);
-
-/**
- * True when a `resourceAvailabilitySummary` entry names an available resource
- * rather than bookkeeping metadata or an unavailable marker.
- */
-export function isAvailableConnectionSourceResource(
-  key: string,
-  value: unknown,
-): boolean {
-  return !CONNECTION_SOURCE_SUMMARY_METADATA_KEYS.has(key)
-    && value !== false
-    && value !== null
-    && value !== undefined;
-}
-
-function countSourceResources(summary: HostedDeviceConnectionSource["resourceAvailabilitySummary"]): number {
-  if (!summary) {
-    return 0;
-  }
-
-  return Object.entries(summary).filter(([key, value]) =>
-    isAvailableConnectionSourceResource(key, value)
-  ).length;
-}
-
-function requiresConnectionSourceReconnect(source: HostedDeviceConnectionSource): boolean {
-  return source.status === "error"
-    && source.lastErrorCode !== null
-    && CONNECTION_SOURCE_RECONNECT_ERROR_CODES.has(source.lastErrorCode);
 }
